+++ sample/src/main/java/com/smileidentity/sample/compose/SettingsScreen.kt
@@ -6,7 +6,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.selection.SelectionContainer
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowForward
+import androidx.compose.material.icons.filled.ArrowForward
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.Divider
 import androidx.compose.material3.Icon
@@ -68,7 +68,7 @@ fun SettingsScreen(
                 ListItem(
                     headlineContent = { Text(stringResource(it.first)) },
                     leadingContent = { Icon(it.second, null) },
-                    trailingContent = { Icon(Icons.AutoMirrored.Filled.ArrowForward, null) },
+                    trailingContent = { Icon(Icons.Filled.ArrowForward, null) },
                     modifier = Modifier.clickable(onClick = it.third),
                 )
                 Divider()